Organize AI Engine kernels
